keep code

+++ frontend/src/views/pages/POS.vue
@@ -14,6 +14,7 @@
                   clearable
                   deletable-chips
                   small-chips
+                  hide-details
                   solo
                 ></v-autocomplete>
               </v-col>
@@ -29,7 +30,6 @@
                 :items="orderItems"
                 :items-per-page="10"
                 class="elevation-4 mx-6"
-                
               >
                 <template v-slot:item.actions="{ item }">
                   <v-icon small class="mr-2" @click="increaseQuantity(item)">
@@ -44,9 +44,8 @@
             </v-row>
           </div>
 
-          
           <v-row class="align-self-end my-0">
-            <v-card  width="100%" class="ma-6">
+            <v-card width="100%" class="ma-6">
               <v-container fluid>
                 <v-row class="mt-2">
                   <v-col cols="6">
@@ -85,6 +84,7 @@
                     small
                     class="rounded-5 mr-lg-5 white--text"
                     color="red"
+                    @click="clearOrderItem"
                   >
                     <v-icon left dark class="mx-0" color="white"
                       >> mdi-cancel
@@ -107,7 +107,7 @@
                     small
                     class="rounded-5 white--text"
                     color="#81c868"
-                    @click="enablePayDialog()"
+                    @click="enablePayDialog"
                   >
                     <v-icon left dark class="mx-0" color="white"
                       >> mdi-cash
@@ -124,11 +124,34 @@
     <v-col lg="8">
       <v-card class="pa-2" outlined tile>
         <v-row class="m-5 pa-5 align-center">
-          <v-text-field label="Search product by name or sku"></v-text-field>
-
-          <v-btn elevation="0" big class="rounded-5 ml-2" color="#81c868">
-            <span class="hidden-sm-and-down">All</span>
-          </v-btn>
+          <v-col cols="4">
+            <v-text-field
+              solo
+              label="Search product by name or sku"
+              clearable
+              width="200px"
+              hide-details
+            ></v-text-field>
+          </v-col>
+          <v-col cols="1">
+            <v-btn
+              elevation="4"
+              big
+              class="rounded-5 ml-2 white--text"
+              color="primary"
+            >
+              <v-icon>mdi-text-box-search-outline</v-icon>
+            </v-btn>
+          </v-col>
+          <v-col cols="7">
+            <v-sheet tile class="py-4 px-1">
+              <v-chip-group multiple active-class="primary--text">
+                <v-chip v-for="cate in categories" :key="cate">
+                  {{ cate }}
+                </v-chip>
+              </v-chip-group>
+            </v-sheet>
+          </v-col>
         </v-row>
         <v-divider class="pa-2"></v-divider>
         <v-row class="m-5 pa-5">
@@ -144,7 +167,11 @@
               width="120"
               src="https://cdn.vuetifyjs.com/images/cards/cooking.png"
             ></v-img> -->
-            <img src="https://cdn.vuetifyjs.com/images/cards/cooking.png" class="item-img" alt="">
+            <img
+              src="https://cdn.vuetifyjs.com/images/cards/cooking.png"
+              class="item-img"
+              alt=""
+            />
 
             <v-card-title
               class="
@@ -174,29 +201,34 @@
         </v-row>
       </v-card>
     </v-col>
-    <DialogDeleteItem
-      v-on:close-delete="closeDelete()"
+    <ConfirmDialog
+      v-on:close-payment-dialog="closeDelete()"
       v-on:delete-item-confirm="deleteItemConfirm()"
       v-bind:showDialog="this.showDialogDelete"
     >
-    </DialogDeleteItem>
-   <PayDialog
-      v-on:close-delete="closePayDialog()"
-      v-bind:enableDialog="this.showPayDialog"
+      <template v-slot:title="">
+        <v-btn>Delete</v-btn>
+      </template>
+    </ConfirmDialog>
+
+    <PayDialog
+      v-on:close-payment-dialog="closePayDialog()"
+      v-on:confirm-payment="confirmPayment"
+      v-bind:show="this.showPayDialog"
     >
-    </PayDialog> 
+    </PayDialog>
   </v-layout>
 </template>
 
 <script>
 import axios from "../../axios";
-import DialogDeleteItem from "../../components/DialogDeleteItem";
-import PayDialog from '../../components/POS/PayDialog.vue'
+import ConfirmDialog from "../../components/ConfirmDialog";
+import PayDialog from "../../components/POS/PayDialog.vue";
 
 export default {
   components: {
-    DialogDeleteItem,
-    PayDialog
+    ConfirmDialog,
+    PayDialog,
   },
   data() {
     return {
@@ -212,215 +244,203 @@ export default {
         { text: "Price", value: "price" },
         { text: "", value: "actions", sortable: false },
       ],
+      customers: [],
       orderItems: [],
       items: [
-        {
-          name: "Bánh chuối tròn",
-          description: null,
-          price: 20000,
-          expired_date: null,
-          image_url: null,
-          quantity: null,
-          category_id: null,
-          stock: null,
-          id: 1,
-          is_active: true,
-        },
-        {
-          name: "Bông lan trứng muối",
-          description: null,
-          price: 24000,
-          expired_date: null,
-          image_url: null,
-          quantity: null,
-          category_id: null,
-          stock: null,
-          id: 2,
-          is_active: true,
-        },
-        {
-          name: "Bánh Thỏi vàng",
-          description: null,
-          price: 14000,
-          expired_date: null,
-          image_url: null,
-          quantity: null,
-          category_id: null,
-          stock: null,
-          id: 3,
-          is_active: true,
-        },
-        {
-          name: "Bánh Mufin sữa",
-          description: null,
-          price: 25000,
-          expired_date: null,
-          image_url: null,
-          quantity: null,
-          category_id: null,
-          stock: null,
-          id: 4,
-          is_active: true,
-        },
-        {
-          name: "Bánh mì kem sữa",
-          description: null,
-          price: 25000,
-          expired_date: null,
-          image_url: null,
-          quantity: null,
-          category_id: null,
-          stock: null,
-          id: 5,
-          is_active: true,
-        },
-        {
-          name: "Bánh Tart phô mai",
-          description: null,
-          price: 20000,
-          expired_date: null,
-          image_url: null,
-          quantity: null,
-          category_id: null,
-          stock: null,
-          id: 6,
-          is_active: true,
-        },
-        {
-          name: "Bánh phô mai bơ tỏi",
-          description: null,
-          price: 20000,
-          expired_date: null,
-          image_url: null,
-          quantity: null,
-          category_id: null,
-          stock: null,
-          id: 7,
-          is_active: true,
-        },
-        {
-          name: "Bánh cade hột gà",
-          description: null,
-          price: 10000,
-          expired_date: null,
-          image_url: null,
-          quantity: null,
-          category_id: null,
-          stock: null,
-          id: 8,
-          is_active: true,
-        },
-        {
-          name: "Bánh Su kem",
-          description: null,
-          price: 15000,
-          expired_date: null,
-          image_url: null,
-          quantity: null,
-          category_id: null,
-          stock: null,
-          id: 9,
-          is_active: true,
-        },
-        {
-          name: "Bánh xúc xích phô mai",
-          description: null,
-          price: 10000,
-          expired_date: null,
-          image_url: null,
-          quantity: null,
-          category_id: null,
-          stock: null,
-          id: 10,
-          is_active: true,
-        },
-        {
-          name: "Bánh mi hạt sen",
-          description: null,
-          price: 30000,
-          expired_date: null,
-          image_url: null,
-          quantity: null,
-          category_id: null,
-          stock: null,
-          id: 11,
-          is_active: true,
-        },
-        {
-          name: "Bánh lột da hạt sen",
-          description: null,
-          price: 20000,
-          expired_date: null,
-          image_url: null,
-          quantity: null,
-          category_id: null,
-          stock: null,
-          id: 12,
-          is_active: true,
-        },
-        {
-          name: "Bánh mì táo đỏ",
-          description: null,
-          price: 25000,
-          expired_date: null,
-          image_url: null,
-          quantity: null,
-          category_id: null,
-          stock: null,
-          id: 13,
-          is_active: true,
-        },
-        {
-          name: "Bánh mini ngọt",
-          description: null,
-          price: 12000,
-          expired_date: null,
-          image_url: null,
-          quantity: null,
-          category_id: null,
-          stock: null,
-          id: 14,
-          is_active: true,
-        },
-        {
-          name: "Bánh bơ tỏi Pháp",
-          description: null,
-          price: 25000,
-          expired_date: null,
-          image_url: null,
-          quantity: null,
-          category_id: null,
-          stock: null,
-          id: 15,
-          is_active: true,
-        },
-        {
-          name: "string",
-          description: "string",
-          price: 0,
-          expired_date: "2021-06-26T12:12:48.117000",
-          image_url: "string",
-          quantity: 0,
-          category_id: null,
-          stock: true,
-          id: 18,
-          is_active: true,
-        },
+        // {
+        //   name: "Bánh chuối tròn",
+        //   description: null,
+        //   price: 20000,
+        //   expired_date: null,
+        //   image_url: null,
+        //   quantity: null,
+        //   category_id: null,
+        //   stock: null,
+        //   id: 1,
+        //   is_active: true,
+        // },
+        // {
+        //   name: "Bông lan trứng muối",
+        //   description: null,
+        //   price: 24000,
+        //   expired_date: null,
+        //   image_url: null,
+        //   quantity: null,
+        //   category_id: null,
+        //   stock: null,
+        //   id: 2,
+        //   is_active: true,
+        // },
+        // {
+        //   name: "Bánh Thỏi vàng",
+        //   description: null,
+        //   price: 14000,
+        //   expired_date: null,
+        //   image_url: null,
+        //   quantity: null,
+        //   category_id: null,
+        //   stock: null,
+        //   id: 3,
+        //   is_active: true,
+        // },
+        // {
+        //   name: "Bánh Mufin sữa",
+        //   description: null,
+        //   price: 25000,
+        //   expired_date: null,
+        //   image_url: null,
+        //   quantity: null,
+        //   category_id: null,
+        //   stock: null,
+        //   id: 4,
+        //   is_active: true,
+        // },
+        // {
+        //   name: "Bánh mì kem sữa",
+        //   description: null,
+        //   price: 25000,
+        //   expired_date: null,
+        //   image_url: null,
+        //   quantity: null,
+        //   category_id: null,
+        //   stock: null,
+        //   id: 5,
+        //   is_active: true,
+        // },
+        // {
+        //   name: "Bánh Tart phô mai",
+        //   description: null,
+        //   price: 20000,
+        //   expired_date: null,
+        //   image_url: null,
+        //   quantity: null,
+        //   category_id: null,
+        //   stock: null,
+        //   id: 6,
+        //   is_active: true,
+        // },
+        // {
+        //   name: "Bánh phô mai bơ tỏi",
+        //   description: null,
+        //   price: 20000,
+        //   expired_date: null,
+        //   image_url: null,
+        //   quantity: null,
+        //   category_id: null,
+        //   stock: null,
+        //   id: 7,
+        //   is_active: true,
+        // },
+        // {
+        //   name: "Bánh cade hột gà",
+        //   description: null,
+        //   price: 10000,
+        //   expired_date: null,
+        //   image_url: null,
+        //   quantity: null,
+        //   category_id: null,
+        //   stock: null,
+        //   id: 8,
+        //   is_active: true,
+        // },
+        // {
+        //   name: "Bánh Su kem",
+        //   description: null,
+        //   price: 15000,
+        //   expired_date: null,
+        //   image_url: null,
+        //   quantity: null,
+        //   category_id: null,
+        //   stock: null,
+        //   id: 9,
+        //   is_active: true,
+        // },
+        // {
+        //   name: "Bánh xúc xích phô mai",
+        //   description: null,
+        //   price: 10000,
+        //   expired_date: null,
+        //   image_url: null,
+        //   quantity: null,
+        //   category_id: null,
+        //   stock: null,
+        //   id: 10,
+        //   is_active: true,
+        // },
+        // {
+        //   name: "Bánh mi hạt sen",
+        //   description: null,
+        //   price: 30000,
+        //   expired_date: null,
+        //   image_url: null,
+        //   quantity: null,
+        //   category_id: null,
+        //   stock: null,
+        //   id: 11,
+        //   is_active: true,
+        // },
+        // {
+        //   name: "Bánh lột da hạt sen",
+        //   description: null,
+        //   price: 20000,
+        //   expired_date: null,
+        //   image_url: null,
+        //   quantity: null,
+        //   category_id: null,
+        //   stock: null,
+        //   id: 12,
+        //   is_active: true,
+        // },
+        // {
+        //   name: "Bánh mì táo đỏ",
+        //   description: null,
+        //   price: 25000,
+        //   expired_date: null,
+        //   image_url: null,
+        //   quantity: null,
+        //   category_id: null,
+        //   stock: null,
+        //   id: 13,
+        //   is_active: true,
+        // },
+        // {
+        //   name: "Bánh mini ngọt",
+        //   description: null,
+        //   price: 12000,
+        //   expired_date: null,
+        //   image_url: null,
+        //   quantity: null,
+        //   category_id: null,
+        //   stock: null,
+        //   id: 14,
+        //   is_active: true,
+        // },
+        // {
+        //   name: "Bánh bơ tỏi Pháp",
+        //   description: null,
+        //   price: 25000,
+        //   expired_date: null,
+        //   image_url: null,
+        //   quantity: null,
+        //   category_id: null,
+        //   stock: null,
+        //   id: 15,
+        //   is_active: true,
+        // },
+        // {
+        //   name: "string",
+        //   description: "string",
+        //   price: 0,
+        //   expired_date: "2021-06-26T12:12:48.117000",
+        //   image_url: "string",
+        //   quantity: 0,
+        //   category_id: null,
+        //   stock: true,
+        //   id: 18,
+        //   is_active: true,
+        // },
       ],
-      customers: [
-        {
-          text: "Walk in customer",
-          value: 0,
-        },
-        {
-          text: "Thien",
-          value: 1,
-        },
-        {
-          text: "Nhan",
-          value: 2,
-        },
-      ],
+      categories: ["Default"],
       customerValue: 0,
       totalItem: 0,
       totalPrice: 0,
@@ -433,14 +453,26 @@ export default {
     };
   },
   mounted() {
-    // this.getListItem();
+    axios.get("/api/v1/items/all?skip=0&limit=100").then((response) => {
+      this.items = response.data;
+    });
+    axios.get("/api/v1/categories").then((response) => {
+      this.categories = response.data.map((item) => item.name);
+    });
+    axios.get("/api/v1/customers").then((response) => {
+      this.customers = response.data.map((item) => ({
+        text: `${item.phone} - ${item.name}`,
+        value: item.phone,
+      }));
+      this.customers.unshift({ text: "Walk in customer", value: -1 });
+      this.customerValue = this.customers[0];
+    });
   },
   methods: {
     initialize() {},
     getListItem() {
       axios.get("/api/v1/items/all?skip=0&limit=100").then((data) => {
         this.items = data.data;
-        console.log(data.data);
       });
     },
     addItemToOrder(item) {
@@ -481,13 +513,28 @@ export default {
       this.grossPrice = totalPrice;
     },
     deleteItem(item) {
-      this.editedIndex = this.orderItems.indexOf(item);
-      this.showDialogDelete = true;
+			console.log(item)
+      this.$swal
+        .fire({
+          title: "Are you sure?",
+          text: "You won't be able to revert this!",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonColor: "#3085d6",
+          cancelButtonColor: "#d33",
+          confirmButtonText: "Yes, delete it!",
+        })
+        .then((result) => {
+          if (result.isConfirmed) {
+            this.editedIndex = this.orderItems.indexOf(item);
+            this.orderItems.splice(this.editedIndex, 1);
+            this.caculateTotalPrice();
+            // this.$swal.fire("Deleted!", "Your file has been deleted.", "success");
+          }
+        });
     },
     deleteItemConfirm() {
-      this.orderItems.splice(this.editedIndex, 1);
       this.closeDelete();
-      this.caculateTotalPrice();
     },
     closeDelete() {
       this.showDialogDelete = false;
@@ -508,13 +555,49 @@ export default {
     },
     enablePayDialog() {
       this.showPayDialog = true;
+      console.log("ahihi");
     },
     closePayDialog() {
       this.showPayDialog = false;
     },
+    confirmPayment(payload) {
+      let orderDetail = {
+        code: new Date().getTime(),
+        user_id: 1,
+        status: true,
+        tax: 0,
+        subtotal: this.totalPrice,
+        paid: payload.amount,
+        change: payload.change,
+        order_items: [
+          {
+            item_id: 0,
+            price: 0,
+            quantity: 0,
+            item_name: "string",
+          },
+        ],
+      };
+
+			orderDetail.order_items = this.orderItems.map(item => ({
+				item_id: item.id,
+				price: item.price,
+				quantity: item.quantity,
+				item_name: item.name
+			}))
+
+			axios.post("/api/v1/orders/", orderDetail).then(response => {
+				console.log(response)
+			})
+
+    },
     blurCustomer() {
+      console.log(this.customerValue);
       if (this.customerValue) return;
       this.customerValue = this.customers[0];
+    },
+    clearOrderItem() {
+      this.orderItems = []
     },
   },
 };
@@ -533,8 +616,12 @@ export default {
   max-width: 120px;
 }
 
-.item-img{
-  height: 120px!important;
-  width: 120px!important;
+.item-img {
+  height: 120px !important;
+  width: 120px !important;
+}
+
+.swal2-popup.swal2-modal.swal2-icon-warning.swal2-show {
+  font-family: "Roboto", sans-serif;
 }
 </style>
